Fixed problem where program wouldn't stop
The program was originally set to run the query until it reaches the 10,001st result (the maximum number allowed by the API). The program now reads the results returned by the query to find the number of results returned, so it can use that as the number to stop running.